refactor(codegen): extrai passes/this_arrow.rs de func.rs
Move lift_arrow_callbacks + struct LiftAcc + impl LiftAcc + helpers
de this/arrow rewriting (~1683 linhas) para passes/this_arrow.rs:
- lift_arrow_callbacks (entry public)
- struct LiftAcc + impl com lift_in_user_fn/lift_in_class/lift_in_body
- arrow_uses_this / stmt_uses_this / expr_uses_this
- arrow_body_to_stmts / rewrite_this_to_under_this
- rewrite_stmt / rewrite_expr (under_this rewrite)
- revert_under_this_inside_arrows / revert_stmt_arrows / revert_expr_arrows
- revert_under_this_in_stmt / revert_under_this_in_expr
- make_this_local / make_slot_assign
- sanitize_for_symbol (movido junto)

Importa de analysis::captures os helpers de promote-to-global.
func.rs: 3763 -> 2070 linhas. Suite verde (11/11 crates).
